add default z3 leaf module class
