+++ deepspeed/utils/__init__.py
@@ -16,7 +16,7 @@ from .tensor_fragment import set_full_hp_param
 from .tensor_fragment import safe_set_full_fp32_param, safe_set_full_optimizer_state
 from .tensor_fragment import safe_get_local_fp32_param, safe_get_local_grad, safe_get_local_optimizer_state
 from .tensor_fragment import safe_set_local_fp32_param, safe_set_local_optimizer_state
-from .tensor_fragment import set_z3_leaf_module, unset_z3_leaf_module, z3_leaf_module
+from .z3_leaf_module import set_z3_leaf_modules, unset_z3_leaf_modules, z3_leaf_module, get_default_z3_leaf_module_classes
 from .mixed_precision_linkage import link_hp_params
 from deepspeed.runtime.dataloader import RepeatingLoader
 from .numa import get_numactl_cmd
